Add an example of using a custom control.
The custom control is just a red button, but all of its logic lives in the
application assembly, rather than the library. This shows you could add
any custom control or logic within user scope.

I've also made it so you don't have to register a control and if
you do not, we make a default factory and store that once.